nextjs-app/autolawyer-mcp/requirements.txt
pypdf
python-docx
chromadb
sentence-transformers
